+++ lib/src/models/eigen/divide_conqour.dart
@@ -65,23 +65,23 @@ class DivideAndConquer {
   }
 
   // Helper function: Combine eigenvectors
-  List<Matrix> combineEigenvectors(List<Matrix> W1, List<Matrix> W2, Matrix P) {
-    int n1 = W1.length;
-    int n2 = W2.length;
+  List<Matrix> combineEigenvectors(List<Matrix> w1, List<Matrix> w2, Matrix P) {
+    int n1 = w1.length;
+    int n2 = w2.length;
 
     List<Matrix> combinedEigenvectors = [];
 
-    // For each eigenvector in W1 and W2, create a new eigenvector as a linear
-    // combination of the eigenvectors in W1 and W2 using the columns of the P matrix
+    // For each eigenvector in w1 and w2, create a new eigenvector as a linear
+    // combination of the eigenvectors in w1 and w2 using the columns of the P matrix
     for (int i = 0; i < n1; i++) {
       Matrix newEigenvector =
-          (P.subMatrix(0, P.rowCount, 0, n1)).transpose() * W1[i];
+          (P.subMatrix(0, P.rowCount, 0, n1)).transpose() * w1[i];
       combinedEigenvectors.add(newEigenvector);
     }
 
     for (int i = 0; i < n2; i++) {
       Matrix newEigenvector =
-          (P.subMatrix(0, P.rowCount, n1, P.columnCount)).transpose() * W2[i];
+          (P.subMatrix(0, P.rowCount, n1, P.columnCount)).transpose() * w2[i];
       combinedEigenvectors.add(newEigenvector);
     }
 
@@ -98,13 +98,13 @@ class DivideAndConquer {
 
     // Step 2: Divide
     int pivot = findPivot(T);
-    Matrix T1 = T.subMatrix(0, 0, pivot + 1, pivot + 1);
-    Matrix T2 = T.subMatrix(pivot + 1, pivot + 1, T.rowCount - pivot - 1,
+    Matrix t1 = T.subMatrix(0, 0, pivot + 1, pivot + 1);
+    Matrix t2 = T.subMatrix(pivot + 1, pivot + 1, T.rowCount - pivot - 1,
         T.columnCount - pivot - 1);
 
     // Step 3: Conquer
-    Eigen eigen1 = solveSmallEigenProblem(T1);
-    Eigen eigen2 = solveSmallEigenProblem(T2);
+    Eigen eigen1 = solveSmallEigenProblem(t1);
+    Eigen eigen2 = solveSmallEigenProblem(t2);
     print(eigen1.values);
 
     // Step 4: Combine